+++ superandes-iteracion2-j.prieto-s.lemus/src/main/java/uniandes/isis2304/superandes/negocio/VOPromocion.java
@@ -18,7 +18,7 @@ public interface VOPromocion {
 
 	public int getCantidad2();
 
-	public int getDescuento();
+	public double getDescuento();
 
 	public long getIdSucural();
